refactor: shared getLabel getColor
