+++ app/components/graphs/HowMuchPerMonthBarChart.tsx
@@ -9,7 +9,7 @@ import {
 import {
   StyledChartContainer,
 } from "../ui/StyledChartContainer";
-import { BarLabelListTopLeft, CustomTick } from "./types";
+import { BarLabelListTopLeft, CustomTick, getLabel, getColor } from "./shared";
 
 type DataInput = {
   category: string;
@@ -79,40 +79,6 @@ const HowMuchPerMonthBarChart: React.FC<StackedBarChartProps> = ({
     },
   ];
   
-  const getLabel = (value: string) => {
-    switch (value) {
-      case "Freehold":
-        return "Freehold";
-      case "Private Rent":
-        return "Private Rent";
-      case "Fairhold - Land Purchase":
-        return "Fairhold /\nLand Purchase";
-      case "Fairhold - Land Rent":
-        return "Fairhold /\nLand Rent";
-      case "Social Rent":
-        return "Social Rent";
-      default:
-        return value;
-    }
-  };
-
-  const getColor = (value: string) => {
-    switch (value) {
-      case "Freehold":
-        return "rgb(var(--not-viable-dark-color-rgb))";
-      case "Private Rent":
-        return "rgb(var(--private-rent-land-color-rgb))";
-      case "Fairhold - Land Purchase":
-        return "rgb(var(--fairhold-equity-color-rgb))";
-      case "Fairhold - Land Rent":
-        return "rgb(var(--fairhold-interest-color-rgb))";
-      case "Social Rent":
-        return "rgb(var(--social-rent-land-color-rgb))";
-      default:
-        return "rgb(var(--fairhold-equity-color-rgb))";
-    }
-  };
-
   return (
     <Card className="h-full w-full">
       <CardContent className="h-full w-full p-0 md:p-4">
