+++ app/components/graphs/ResaleValueLineChart.tsx
@@ -9,7 +9,7 @@ import {
 } from "../ui/StyledChartContainer";
 import { formatValue } from "@/app/lib/format";
 import { MaintenanceLevel } from "@/app/models/constants";
-import { CustomLabelListContentProps } from "./types";
+import { CustomLabelListContentProps } from "./shared";
 
 type CustomTooltipProps = TooltipProps<number, string> & {
   payload?: Array<{
